% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-and-read-data.R
\name{download_and_read_data}
\alias{download_and_read_data}
\title{Download and read review data}
\usage{
download_and_read_data(save_location = NULL)
}
\arguments{
\item{save_location}{character. The contents of the downloaded zip file will
be uncompressed and saved here. If save_location does not exist, it will be
created, as well as any missing parent directories.}
}
\value{
Tibble of parsed data
}
\description{
Download and read review data
}
