% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download_data}
\alias{download_data}
\title{Download and uncompress a zipped data file}
\usage{
download_data(save_location = getwd())
}
\arguments{
\item{save_location}{character. The contents of the downloaded zip file will
be uncompressed and saved here. If save_location does not exist, it will be
created, as well as any missing parent directories.}
}
\description{
Given the URL of a .zip file, this function will download it and decompress
it to a given location.
}
\details{
This function was intended to download and uncompress the data located at
https://archive.ics.uci.edu/ml/datasets/Sentiment+Labelled+Sentences
It is not a general function to download arbitrary data, but may be a
useful template.
}
