% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-artefact.R
\name{export_artefact}
\alias{export_artefact}
\title{Export a model artefact to inst/artefacts}
\usage{
export_artefact(artefact, name = NULL)
}
\arguments{
\item{artefact}{Object that will be exported as an RDS}

\item{name}{File name to use, without the directory. If not provided, will
attempt to save as "artefact.rds", although this will fail for piped objects.}
}
\value{
Returns the artefact, invisibly.
}
\description{
A model artefact could be a model itself, or an object used in preprocessing
(like a recipe). This function saves artefacts to the inst/artefacts 
directory of the package, but it could just as easily export to a remote
resource like cloud storage.
}
