% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-artefact.R
\name{import_artefact}
\alias{import_artefact}
\title{Import a model artefact from inst/artefacts}
\usage{
import_artefact(artefact_name)
}
\arguments{
\item{artefact_name}{RDS object that will be imported}
}
\value{
The object parsed by readr
}
\description{
A model artefact could be a model itself, or an object used in preprocessing
(like a recipe). This function reads artefacts from the inst/artefacts 
directory of the package, but it could just as easily import from a remote
resource like cloud storage.
}
