% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-execution-plan.R
\name{model_execution_plan}
\alias{model_execution_plan}
\title{drake plan for model execution}
\usage{
model_execution_plan()
}
\value{
A plan to be run with drake::make()
}
\description{
drake plan for model execution
}
