% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-training-plan.R
\name{model_training_plan}
\alias{model_training_plan}
\title{drake plan for model training}
\usage{
model_training_plan()
}
\value{
A plan to be run with drake::make()
}
\description{
drake plan for model training
}
