% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-data-to-be-scored.R
\name{new_data_to_be_scored}
\alias{new_data_to_be_scored}
\title{Retrieve new data that needs to be scored}
\usage{
new_data_to_be_scored()
}
\value{
Data frame with two char columns: review and sentiment ("good"/"bad")
}
\description{
This function is meant to replicate a function that will retrieve new data
that needs to be scored. This is a batch execution, rather than an on-demand
API that might be stood up by the plumber package. In practice, the return
is a fixed sample of 10 reviews.
}
