% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-review-file.R
\name{read_review_file}
\alias{read_review_file}
\title{Read a review source data file into a data frame}
\usage{
read_review_file(file_path)
}
\arguments{
\item{file_path}{character. A file path to a source .txt data file}
}
\description{
This function is specifically intended to parse the three data files in the
sentiment analysis data set
(https://archive.ics.uci.edu/ml/datasets/Sentiment+Labelled+Sentences).
}
