% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem-tokeniser.R
\name{stem_tokeniser}
\alias{stem_tokeniser}
\title{Stem and tokenise a sentence}
\usage{
stem_tokeniser(x)
}
\arguments{
\item{x}{A character or vector of characters, usually sentences, paragraphs
or similar pieces of natural language.}
}
\value{
A vector of tokens
}
\description{
Stem and tokenise a sentence
}
\examples{
stem_tokeniser("information informed informing informs")

}
