% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-predictions.R
\name{submit_predictions}
\alias{submit_predictions}
\title{Simulate submission of output data}
\usage{
submit_predictions(predictions)
}
\arguments{
\item{predictions}{Predictions to be submitted. Format is not important.}
}
\value{
Returns output, invisibly
}
\description{
This is a dummy function that simulates the submission of newly scored 
model results to an external platform. In actual fact, this function does
nothing, but sleeps for 5 seconds to simulate an upload process.
}
