% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-preprocessor.R
\name{text_preprocessor}
\alias{text_preprocessor}
\title{Create a pruned vocabulary from a token iterator}
\usage{
text_preprocessor(x)
}
\arguments{
\item{x}{Character. Text to be processed.}
}
\value{
The same character after processing.
}
\description{
This function creates a vocabulary from a vector of documents. A
vocabulary defines the domain of a natural language processing problem.
Vocabularies are often used to create vectorisers, which allow novel pieces
of text to be mapped to a vocabulary defined by a training set. To exclude
frequently and infrequently occurring tokens, the vocabulary is often
trimmed. This reduces the dimension of the problem to decrease training time
and the potential for overfitting.
}
