% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-model.R
\name{validate_model}
\alias{validate_model}
\title{Assert that model makes sensible predictions}
\usage{
validate_model(random_forest, vectoriser, tfidf = NULL)
}
\arguments{
\item{random_forest}{A model created with the randomForest package.}

\item{vectoriser}{A vectoriser constructed with the text2vec package.}

\item{tfidf}{A tfidf object constructed with the text2vec package.
If no tfidf is NULL, then weighting will not be applied.}
}
\value{
The MD5 hash of the random_forest object
}
\description{
This function tests model predictions against a small sample of inputs with
predictable outputs. For example, a sentiment analysis model that returns
a sentiment of "bad" for "love" should not be used. We can also demand a
minimum level of accuracy according to one or more metrics. Any failed 
assertions will raise an error. This validation approach could also be used
to perform rudimentary checks for model bias (in the ethics sense).
}
