% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-predictions.R
\name{validate_predictions}
\alias{validate_predictions}
\title{Assert that model predictions are sensible}
\usage{
validate_predictions(predictions)
}
\arguments{
\item{predictions}{A vector of model predictions}
}
\value{
The MD5 hash of the predictions
}
\description{
This function tests that model predictions make sense. It is intended to be
run after new data has been scored, but before it has been submitted. We
might ensure that outputs are sensible, within an expected range, or do not
contain sensitive data. Failed assertions will raise an error. Upon success,
the md5 hash of the predictions will be returned.
}
