% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-and-read-data.R
\name{download_and_read_data}
\alias{download_and_read_data}
\title{Download and parse review data}
\usage{
download_and_read_data(data_source_url)
}
\arguments{
\item{data_source_url}{Character.}
}
\value{
Tibble of parsed data
}
\description{
Given the URL of a .zip file, this function will download it and decompress
it to a given location.
}
\details{
This function was intended to download and uncompress the data located at
https://archive.ics.uci.edu/ml/datasets/Sentiment+Labelled+Sentences
It is not a general function to download arbitrary data, but may be a
useful template.
}
